update .gitignore to exclude new files related to Gemini and temporary credentials
