+++ .gitignore
@@ -48,3 +48,4 @@ x64*
 gha-creds-*.json
 GEMINI.md
 .kilocodemodes
+compile_commands.json
