+++ CMakeLists.txt
@@ -74,6 +74,7 @@ target_link_libraries(vkml-compiler
  	# "MLIRTensorMeshShardingExtensions" 
  	# "MLIRTensorTilingInterfaceImpl"
 	"MLIRTensorToLinalg" 
+	"MLIRMLProgramTransforms"
 # 	"MLIRTensorTransformOps"
 # 	"MLIRTensorToSPIRV" 
 # 	"MLIRTensorTransforms"
@@ -83,15 +84,15 @@ target_link_libraries(vkml-compiler
 	"MLIRMLProgramDialect"
 	"MLIRTosaToMLProgram"
 	"MLIRTosaToLinalg" 
-# 	"MLIRTosaToArith"
+	"MLIRTosaToArith"
 	"MLIRTosaToTensor"
 	"MLIRTosaToSCF"
 # 	"MLIRTosaShardingInterfaceImpl"
 	"MLIRGPUDialect"
 	"MLIRGPUTransforms"
+	"MLIRSCFToGPU"
 # 	"MLIRSPIRVDialect"
 # 	"MLIRAffineToStandard"
-# 	"MLIRSCFToGPU"
 # 	"MLIRBufferizationToMemRef"
 )
